# Improvements on OriginalSource and ReplaceSource