--- a/lib/OriginalSource.js
+++ b/lib/OriginalSource.js
@@ -9,32 +9,10 @@ var SourceMapConsumer = require("source-map").SourceMapConsumer;
 var SourceListMap = require("source-list-map").SourceListMap;
 var Source = require("./Source");
 
-function isSplitter(c) {
-	switch(c) {
-		case 10: // \n
-		case 13: // \r
-		case 59: // ;
-		case 123: // {
-		case 125: // }
-			return true;
-	}
-	return false;
-}
+var SPLIT_REGEX = /(?!$)[^\n\r;{}]*[\n\r;{}]*/g;
 
 function _splitCode(code) {
-	var result = [];
-	var i = 0;
-	var j = 0;
-	for(; i < code.length; i++) {
-		if(isSplitter(code.charCodeAt(i))) {
-			while(isSplitter(code.charCodeAt(++i)));
-			result.push(code.substring(j, i));
-			j = i;
-		}
-	}
-	if(j < code.length)
-		result.push(code.substr(j));
-	return result;
+	return code.match(SPLIT_REGEX) || [];
 }
 
 class OriginalSource extends Source {