--- a/lib/ReplaceSource.js
+++ b/lib/ReplaceSource.js
@@ -75,7 +75,7 @@ class ReplaceSource extends Source {
 		if(this.replacements.length === 0) {
 			return node;
 		}
-		this.replacements.sort(sortReplacementsAscending);
+		this._sortReplacements();
 		var replace = new ReplacementEnumerator(this.replacements);
 		var output = [];
 		var position = 0;
@@ -258,29 +258,14 @@ class ReplaceSource extends Source {
 	}
 }
 
-function sortReplacementsAscending(a, b) {
-	var diff = a[1] - b[1]; // end
-	if(diff !== 0)
-		return diff;
-	diff = a[0] - b[0]; // start
-	if(diff !== 0)
-		return diff;
-	return a[3] - b[3]; // insert order
-}
-
 class ReplacementEnumerator {
 	constructor(replacements) {
-		this.emit = true;
-		this.done = !replacements || replacements.length === 0;
-		this.index = 0;
-		this.replacements = replacements;
-		if(!this.done) {
-			// Set initial start position in case .header is not called
-			var repl = replacements[0];
-			this.position = Math.floor(repl[0]);
-			if(this.position < 0)
-				this.position = 0;
-		}
+		this.replacements = replacements || [];
+		this.index = this.replacements.length;
+		this.done = false;
+		this.emit = false;
+		// Set initial start position
+		this.next();
 	}
 
 	next() {
@@ -294,8 +279,8 @@ class ReplacementEnumerator {
 			this.value = repl[2];
 		} else {
 			// End point found. start emitting. set position to find next start
-			this.index++;
-			if(this.index >= this.replacements.length) {
+			this.index--;
+			if(this.index < 0) {
 				this.done = true;
 			} else {
 				var nextRepl = this.replacements[this.index];
@@ -312,9 +297,16 @@ class ReplacementEnumerator {
 	footer() {
 		if(!this.done && !this.emit)
 			this.next(); // If we finished _replaceInNode mid emit we advance to next entry
-		return this.done ? [] : this.replacements.slice(this.index).map(function(repl) {
-			return repl[2];
-		}).join("");
+		if(this.done) {
+			return [];
+		} else {
+			var resultStr = "";
+			for(var i = this.index; i >= 0; i--) {
+				var repl = this.replacements[i];
+				resultStr += repl[2];
+			}
+			return resultStr;
+		}
 	}
 }