**OriginalSource**
Use `str.match(regexp)` in `_splitCode`, which is faster than string traversal with for loop

**RepalceSource**
Make `ReplacementEnumerator` take the replacements in descending order and remove redundant code.