@jevancc 2018-09-05
@evilebottnawi it should be fine now. The small decrease in test coverage is because of the removing of redundant code which was previously hit in the test.