@ooflorent 2018-09-07
> which is faster than string traversal with for loop

Can you create a gist with a benchmark please?

Why did you change the iteration order? I found the proposed solution harder to understand. Could you elaborate a bit more the motiviation?