@jevancc 2018-09-07
@ooflorent 
> Can you create a gist with a benchmark please?
* Node.js version: 10.8.0

|  String length | Portion of Splitter |**Old split** | **New split** |
|----------------|---------------------|--------------|---------------|
| 1,000,000 | 30% | 49.42ms | 49.07ms |
| 1,000,000 | 20% | 38.33ms | 36.21ms |
| 1,000,000 | 10% | 21.83ms | 18.40ms |
| 1,000,000 |  5% | 13.76ms | 7.41ms |

More details: [Benchmark gist](https://gist.github.com/jevancc/bfd43a613f350b46d0b2b48334089584)

> I found the proposed solution harder to understand. Could you elaborate a bit more the motiviation?

My motivation is very simple: other functions in `ReplaceSource` (`.source()`, `.listMap()`) all take the replacements in descending order and use the same sorting method `._sortReplacements()`.

Since this change is not that complicated, it is better not to maintain another sorting function and re-sort the replacements in different order everytime when calling `.node()`.